add naive account indexing
