+++ frontend/src/components/account/avatar/AccountAvatar.tsx
@@ -1,5 +1,23 @@
-import { FlowUtils } from "../../../utils/flow-utils";
 import React, { ReactElement } from "react";
+
+import service from "../../../assets/images/avatars/service.svg";
+
+import avatar1 from "../../../assets/images/avatars/1.jpg";
+import avatar2 from "../../../assets/images/avatars/2.jpg";
+import avatar3 from "../../../assets/images/avatars/3.jpg";
+import avatar4 from "../../../assets/images/avatars/4.jpg";
+import avatar5 from "../../../assets/images/avatars/5.jpg";
+import avatar6 from "../../../assets/images/avatars/6.jpg";
+import avatar7 from "../../../assets/images/avatars/7.jpg";
+import avatar8 from "../../../assets/images/avatars/8.jpg";
+import avatar9 from "../../../assets/images/avatars/9.jpg";
+import avatar10 from "../../../assets/images/avatars/10.jpg";
+import avatar11 from "../../../assets/images/avatars/11.jpg";
+import avatar12 from "../../../assets/images/avatars/12.jpg";
+import avatar13 from "../../../assets/images/avatars/13.jpg";
+import avatar14 from "../../../assets/images/avatars/14.jpg";
+import avatar15 from "../../../assets/images/avatars/15.jpg";
+import avatar16 from "../../../assets/images/avatars/16.jpg";
 
 type AccountAvatarProps = {
   size?: number;
@@ -22,7 +40,44 @@ export function AccountAvatar({
         width: size,
       }}
       alt={address}
-      src={FlowUtils.getUserAvatarUrl(address)}
+      src={getUserAvatarUrl(address)}
     />
   );
+}
+
+const avatarUrls = [
+  avatar1,
+  avatar2,
+  avatar3,
+  avatar4,
+  avatar5,
+  avatar6,
+  avatar7,
+  avatar8,
+  avatar9,
+  avatar10,
+  avatar11,
+  avatar12,
+  avatar13,
+  avatar14,
+  avatar15,
+  avatar16,
+];
+
+// https://github.com/onflow/flow-go/blob/master/model/flow/chain.go#L98-L113
+function getUserAvatarUrl(address: string): string {
+  const isServiceAccount = [
+    "0xf8d6e0586b0a20c7",
+    "0x0000000000000001", // When using monotonic addresses setting
+  ].includes(address);
+
+  if (isServiceAccount) {
+    return service;
+  }
+
+  const normalizedDecAddress = Math.round(parseInt(address, 16) / 100000);
+
+  const avatarIndex = normalizedDecAddress % avatarUrls.length;
+
+  return avatarUrls[avatarIndex];
 }
