+++ frontend/src/react-app-env.d.ts
@@ -17,3 +17,5 @@ declare module "*.module.scss" {
   const classes: { readonly [key: string]: string };
   export default classes;
 }
+
+declare module "*.jpg";
